refactor: JPA 클래스 파일들을 infra 레이어에서 domain 레이어로 변경
- review : JPA 클래스들을 infra 레이어에 둘 때는, entity와 domain을 분리 설계한 pojo 방식
