+++ qrust-domain/src/main/java/com/qrust/qrcode/domain/repository/QrCodeImageRepository.java
@@ -1,4 +1,4 @@
-package com.qrust.qrcode.infrastructure.jpa;
+package com.qrust.qrcode.domain.repository;
 
 import com.qrust.qrcode.domain.entity.QrCodeImage;
 import org.springframework.data.jpa.repository.JpaRepository;
