+++ qrust-domain/src/main/java/com/qrust/qrcode/domain/repository/QrCodeRepository.java
@@ -1,4 +1,4 @@
-package com.qrust.qrcode.infrastructure.jpa;
+package com.qrust.qrcode.domain.repository;
 
 import com.qrust.qrcode.domain.entity.QrCode;
 import org.springframework.data.jpa.repository.JpaRepository;
